+++ firmware/hdl/packages/gt_mp7_core_pkg.vhd
@@ -1,14 +1,15 @@
 
 -- actual versions:
 -- use "FRAME_VERSION" as mp7_ugt release fw version (used for tag name).
--- mp7_ugt (=FRAME_VERSION): v1.20.3
+-- mp7_ugt (=FRAME_VERSION): v1.21.1
 
 -- use "GT_VERSION" as mp7_ugt release fw version (used for tag name).
--- gt: v1.20.3
+-- gt: v1.21.1
 -- frame: v1.3.1 (see frame.vhd)
 -- gtl: v1.17.3 (see gtl_module_tpl.vhd)
--- fdl: v1.3.8 (see fdl_module.vhd)
+-- fdl: v1.4.1 (see fdl_module.vhd)
 
+-- HB 2022-05-31: v1.21.0 - Signal start_sync_bc0_int used for reset of prescale counter (instead of begin_lumi_section). Removed sres signals for counters in FDL, not used anymore.
 -- HB 2021-05-31: v1.20.2 - Updated script run_simulation_questa.py (removed UGT_GITLAB_PWD)
 -- HB 2021-04-25: v1.20.1 - Updated and added scripts.
 -- HB 2021-02-03: v1.20.0 - "GT_VERSION" is (used in frame_module_info.vhd) in register of "FRAME_VERSION". "FRAME_VERSION" is now in register of "SVN_REVISION_NUMBER".
@@ -23,7 +24,6 @@
 -- v1.15.3: Added python script runSynthWorkFlow.py
 -- v1.15.2: Added tcl script for "manualy" bit file generation (after timing errors)
 --
-
 -- gtl history:
 -- HB 2022-09-02: v1.17.3: cleaned up.
 -- HB 2022-02-16: v1.17.2: Bug fixed in comb_conditions.vhd.
@@ -76,8 +76,8 @@ package gt_mp7_core_pkg is
 -- ==================================================================================================
 -- GT firmware version
     constant GT_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant GT_MINOR_VERSION      : integer range 0 to 255 := 20;
-    constant GT_REV_VERSION        : integer range 0 to 255 := 3;
+    constant GT_MINOR_VERSION      : integer range 0 to 255 := 21;
+    constant GT_REV_VERSION        : integer range 0 to 255 := 1;
 	constant GT_VERSION : std_logic_vector(31 downto 0) := X"00" &
            std_logic_vector(to_unsigned(GT_MAJOR_VERSION, 8)) &
            std_logic_vector(to_unsigned(GT_MINOR_VERSION, 8)) &
@@ -92,8 +92,9 @@ package gt_mp7_core_pkg is
     constant GTL_FW_REV_VERSION        : integer range 0 to 255 := 3;
 -- FDL firmware version
     constant FDL_FW_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant FDL_FW_MINOR_VERSION      : integer range 0 to 255 := 3;
-    constant FDL_FW_REV_VERSION        : integer range 0 to 255 := 8;
+    constant FDL_FW_MINOR_VERSION      : integer range 0 to 255 := 4;
+    constant FDL_FW_REV_VERSION        : integer range 0 to 255 := 1;
+
 -- ==================================================================================================
 
     constant BUNCHES_PER_ORBIT : natural range 3564 to 3564 := LHC_BUNCH_COUNT;